refactor(git-graph): split GitGraphPane into commit-list feature and components
intent(git-graph): 1300 行超の GitGraphPane を責務ごとに分割する
decision(commit-list): グラフ本体を commit-list feature に切り出し、barrel は CommitGraphList のみ公開して行・SVG・レーン算出を隠蔽
rejected(graph-subfeature): graph 描画を独立 feature にしない。layout が行描画と SVG の共有 spine で barrel から隠せず全 export になるため commit-list 内の component/module に留める
rejected(composable-extract): 行数削減目的の composable 退避を却下。composable は複数 component 共有のための道具
constraint(scroll-to-head): data 所有の pane と DOM 所有の list を store の scrollToHeadToken signal で疎結合にする
learned(commit-row): 右クリックの e.currentTarget は pointerup 待機中に null 化するため anchor / 座標を contextmenu 発火時に同期 snapshot する
