+++ apps/renderer/src/features/git-graph/features/commit-context-menu/index.ts
@@ -1,0 +1,2 @@
+export { default as CommitContextMenu } from "./CommitContextMenu.vue";
+export { useCommitContextMenuTrigger } from "./useCommitContextMenuTrigger";
