+++ apps/renderer/src/features/git-graph/features/commit-list/CommitGraphList.vue
@@ -1,0 +1,287 @@
+<doc lang="md">
+commit graph 本体。commits からレーンレイアウトを算出し、Working Tree 行・SVG overlay・commit 行を
+組み合わせて描く。キーボードナビ・スクロール・行クリック選択を担う。
+
+## layout は行と SVG の共有 spine
+
+行の v-for と SVG の dot は同じ layout を辿り、同じ行高で重なる。これで gap 行挿入を含め dot と行が
+常に整合する。片方だけに layout を持たせず 1 つを両方へ配る。
+
+## HEAD スクロール
+
+data 取得側 (親) が出す scroll 要求 signal を watch して実スクロールする。DOM 反映後に走らせるため
+flush: "post"。列の縦揃えは Working Tree 行と commit 行が同じ `--graph-cols` を引くことで保つ。
+</doc>
+
+<script setup lang="ts">
+import type { GitPullRequest } from "@gozd/proto";
+import { storeToRefs } from "pinia";
+import { computed, ref, watch } from "vue";
+import { computeStatusIcons, UNCOMMITTED_HASH, useGitStatusStore } from "../../../worktree";
+import { linkifyCommitMessage } from "../../linkifyCommitMessage";
+import { useGitGraphStore } from "../../useGitGraphStore";
+import CommitRow from "./CommitRow.vue";
+import GapRow from "./GapRow.vue";
+import { ROW_HEIGHT, graphColumnWidth as calcGraphColumnWidth } from "./graphGeometry";
+import { computeGraphLayout } from "./graphLayout";
+import { computeOutOfSyncBranches } from "./graphRefs";
+import GraphSvgOverlay from "./GraphSvgOverlay.vue";
+import WorkingTreeRow from "./WorkingTreeRow.vue";
+
+const props = defineProps<{
+  /** デフォルトブランチ名 (ref を default タイプに分類する)。loadLog 結果由来で親が渡す */
+  defaultBranch?: string;
+  /** GitHub repo base URL (コミットメッセージ `#N` リンク化)。親が repo identity から導出 */
+  issueLinkBaseUrl?: string;
+  /** ブランチ名 → PR のマップ。親が prListStore から渡す */
+  prByBranch: Map<string, GitPullRequest>;
+}>();
+
+const emit = defineEmits<{
+  commitContextmenu: [payload: { hash: string; anchorEl: HTMLElement; x: number; y: number }];
+}>();
+
+const gitGraphStore = useGitGraphStore();
+const gitStatusStore = useGitStatusStore();
+const { commits } = storeToRefs(gitGraphStore);
+const { gitStatuses, workingTreeMtime } = storeToRefs(gitStatusStore);
+
+/** commits + HEAD からレーンレイアウトを算出。commits 変化で自動再計算する。 */
+const layout = computed(() =>
+  computeGraphLayout(commits.value, { headHash: gitGraphStore.headHash }),
+);
+
+/** 変更ファイル数 (Working Tree 行) */
+const uncommittedChangeCount = computed(() => Object.keys(gitStatuses.value).length);
+/** 変更をアイコン付きカウントに変換 (Working Tree 行) */
+const statusIcons = computed(() => computeStatusIcons(gitStatuses.value));
+
+/** HEAD が指すカレントブランチ名 (ref を current タイプに分類する) */
+const currentBranch = computed(() => gitGraphStore.currentBranch);
+
+/** ローカルとリモートが別コミットに分かれているブランチ名の集合 */
+const outOfSyncBranches = computed(() => computeOutOfSyncBranches(commits.value));
+
+/**
+ * refs に "HEAD" を持つ表示中ノードとその行番号。Working Tree 行の色・HEAD スクロールの導出元。
+ */
+const headNode = computed(() => {
+  const index = layout.value.nodes.findIndex((n) => n.commit.refs.includes("HEAD"));
+  if (index === -1) return undefined;
+  return { node: layout.value.nodes[index], index };
+});
+
+/** HEAD ノードの色インデックス。Working Tree のドット/接続線を HEAD と同色に揃える。不在時 0。 */
+const headColor = computed(() => headNode.value?.node.color ?? 0);
+
+/** Graph 列の幅。右側は HEAD marker 分の余白を確保する。 */
+const graphColumnWidth = computed(() => calcGraphColumnWidth(layout.value.maxLanes));
+
+/** グラフ全体の高さ (行数 × 行高)。scroll 領域の minHeight に使う。 */
+const svgHeight = computed(() => layout.value.nodes.length * ROW_HEIGHT);
+
+/** 表示中の各 commit の subject を linkify した segments を hash で引ける map。
+ * template から関数呼び出しすると毎 render で全 commit 分 matchAll が走るため computed に閉じる。 */
+const commitMessageSegmentsByHash = computed(() => {
+  const baseUrl = props.issueLinkBaseUrl;
+  const map = new Map<string, ReturnType<typeof linkifyCommitMessage>>();
+  for (const node of layout.value.nodes) {
+    map.set(node.commit.hash, linkifyCommitMessage(node.commit.message, baseUrl));
+  }
+  return map;
+});
+
+// --- スクロール / キーボードナビ ---
+
+const scrollContainer = ref<HTMLElement | null>(null);
+
+/** 現在選択中のノードの **行 (node) インデックス**。範囲選択中は compareHash（移動端）を返す。
+ * gap 行挿入で commit index と node index がずれるため layout.nodes 上の位置を引く。未選択 / WT は -1。 */
+function selectedIndex(): number {
+  const { compareHash } = gitGraphStore;
+  const hash = compareHash ?? gitGraphStore.selectedHash;
+  return layout.value.nodes.findIndex((n) => !n.gap && n.commit.hash === hash);
+}
+
+/** index から dir 方向へ gap 行を読み飛ばした最初の非 gap 行を返す。範囲外はそのまま返し caller が clamp する。 */
+function skipGapRows(index: number, dir: number): number {
+  const nodes = layout.value.nodes;
+  let i = index;
+  while (i >= 0 && i < nodes.length && nodes[i]?.gap) i += dir;
+  return i;
+}
+
+/** 選択を Uncommitted Changes に戻し、HEAD コミット付近にスクロール */
+function scrollHeadIntoView() {
+  const head = headNode.value;
+  if (head === undefined) return;
+  gitGraphStore.resetSelection();
+  scrollToCenter(head.index);
+}
+
+/** 指定行をビューポート中央にスクロール */
+function scrollToCenter(index: number) {
+  const container = scrollContainer.value;
+  if (!container) return;
+  const rowCenter = index * ROW_HEIGHT + ROW_HEIGHT / 2;
+  container.scrollTop = rowCenter - container.clientHeight / 2;
+}
+
+/** 選択行をビューポート内にスクロール */
+function scrollToIndex(index: number) {
+  const container = scrollContainer.value;
+  if (!container) return;
+  const rowTop = index * ROW_HEIGHT;
+  const rowBottom = rowTop + ROW_HEIGHT;
+  if (rowTop < container.scrollTop) {
+    container.scrollTop = rowTop;
+  } else if (rowBottom > container.scrollTop + container.clientHeight) {
+    container.scrollTop = rowBottom - container.clientHeight;
+  }
+}
+
+/** ビューポートに収まる行数 */
+function pageSize(): number {
+  const container = scrollContainer.value;
+  if (!container) return 1;
+  return Math.max(1, Math.floor(container.clientHeight / ROW_HEIGHT));
+}
+
+// 親が HEAD スクロールを要求したとき (worktree 切替 / HEAD 移動 / Scroll to HEAD ボタン) 実行する。
+// flush: "post" で DOM 反映後に走らせ、commits 変化直後 (行 / minHeight 未反映) の誤スクロールを避ける。
+watch(
+  () => gitGraphStore.scrollToHeadToken,
+  () => scrollHeadIntoView(),
+  { flush: "post" },
+);
+
+function onKeydown(e: KeyboardEvent) {
+  const nodes = layout.value.nodes;
+  if (nodes.length === 0) return;
+
+  const isUp = e.key === "ArrowUp" || e.key === "PageUp";
+  const isDown = e.key === "ArrowDown" || e.key === "PageDown";
+  if (!isUp && !isDown) return;
+  e.preventDefault();
+
+  const isPage = e.key === "PageUp" || e.key === "PageDown";
+  const step = isPage ? pageSize() : 1;
+  const current = selectedIndex();
+
+  // 移動端が Working Tree にある場合（単一選択 or 範囲選択の compareHash が UNCOMMITTED_HASH）
+  if (current === -1) {
+    if (isUp) {
+      scrollToIndex(0);
+      return;
+    }
+    const next = skipGapRows(Math.min(step - 1, nodes.length - 1), 1);
+    if (next >= nodes.length) return;
+    const hash = nodes[next].commit.hash;
+    if (e.shiftKey) {
+      gitGraphStore.selectCompare(hash);
+    } else {
+      gitGraphStore.select(hash);
+    }
+    scrollToIndex(next);
+    return;
+  }
+
+  // コミット行先頭付近で上移動 → Working Tree へ
+  if (isUp && current !== -1 && current - step < 0) {
+    if (e.shiftKey) {
+      gitGraphStore.selectCompare(UNCOMMITTED_HASH);
+    } else {
+      gitGraphStore.select(UNCOMMITTED_HASH);
+    }
+    return;
+  }
+
+  let next: number;
+
+  if (isUp) {
+    next = skipGapRows(current - step, -1);
+    // gap を上に抜けて Working Tree 域に出たら Working Tree 選択へ倒す。
+    if (next < 0) {
+      if (e.shiftKey) {
+        gitGraphStore.selectCompare(UNCOMMITTED_HASH);
+      } else {
+        gitGraphStore.select(UNCOMMITTED_HASH);
+      }
+      return;
+    }
+  } else {
+    next = skipGapRows(Math.min(nodes.length - 1, current + step), 1);
+    if (next >= nodes.length) return;
+  }
+
+  if (e.shiftKey) {
+    gitGraphStore.selectCompare(nodes[next].commit.hash);
+  } else {
+    gitGraphStore.select(nodes[next].commit.hash);
+  }
+  scrollToIndex(next);
+}
+
+function onRowClick(hash: string, e: MouseEvent) {
+  // macOS WebKit は control+click を button=0 の click として dispatch する (webkit bugzilla 52174)。
+  // contextmenu と一緒に通常 click も発火するため、control+click は選択変更に倒さず contextmenu 経路に委譲する。
+  if (e.ctrlKey) return;
+  if (e.shiftKey) {
+    gitGraphStore.selectCompare(hash);
+  } else {
+    gitGraphStore.select(hash);
+  }
+}
+</script>
+
+<template>
+  <!-- `--graph-cols` を SSOT として持ち、Working Tree 行 / commit 行の双方が
+       grid-template-columns: var(--graph-cols) で同じ列幅を引く。sticky 行と scroll 配下の
+       commit 行は別 DOM container にあるため subgrid では繋げない。
+       description 列は minmax(0, 1fr) — 素の 1fr だと min-content を要求して truncate と両立しない。 -->
+  <div
+    class="flex min-w-0 flex-1 flex-col outline-none"
+    :style="{ '--graph-cols': `${graphColumnWidth}px minmax(0, 1fr) 7rem 7rem 4rem` }"
+    tabindex="0"
+    @keydown="onKeydown"
+  >
+    <!-- スクロール可能なコミットリスト。Working Tree 行はこの中に sticky で置き、commit 行と
+         同じ effective 幅を共有させて column を揃える。 -->
+    <div ref="scrollContainer" class="min-h-0 flex-1 overflow-auto">
+      <WorkingTreeRow
+        :graph-column-width="graphColumnWidth"
+        :head-color="headColor"
+        :change-count="uncommittedChangeCount"
+        :status-icons="statusIcons"
+        :mtime="workingTreeMtime"
+        @row-click="onRowClick(UNCOMMITTED_HASH, $event)"
+      />
+
+      <div class="relative" :style="{ minHeight: `${svgHeight}px` }">
+        <GraphSvgOverlay
+          :layout="layout"
+          :graph-column-width="graphColumnWidth"
+          :head-color="headColor"
+        />
+
+        <!-- Commit table rows: row には `position` を付けない (非 positioned = stacking layer 3)。
+             兄弟の overlay SVG は positioned absolute (layer 6) なので、CSS の painting order により
+             z-index を一切使わずに SVG が row 背景の上に描かれる。 -->
+        <template v-for="(node, row) in layout.nodes" :key="`row-${row}`">
+          <GapRow v-if="node.gap" />
+          <CommitRow
+            v-else
+            :node="node"
+            :current-branch="currentBranch"
+            :default-branch="defaultBranch"
+            :out-of-sync-branches="outOfSyncBranches"
+            :pr-by-branch="prByBranch"
+            :segments="commitMessageSegmentsByHash.get(node.commit.hash) ?? []"
+            @row-click="onRowClick"
+            @row-contextmenu="emit('commitContextmenu', $event)"
+          />
+        </template>
+      </div>
+    </div>
+  </div>
+</template>
