+++ apps/renderer/src/features/git-graph/features/commit-list/displayRef.ts
