+++ apps/renderer/src/features/git-graph/features/commit-list/graphColors.ts
