+++ apps/renderer/src/features/git-graph/features/commit-list/index.ts
@@ -1,0 +1,1 @@
+export { default as CommitGraphList } from "./CommitGraphList.vue";
